[Move]SendSoundEffectを作成し、関連メソッドを移動
