+++ src/main/scala/com/github/unchama/seichiassist/listener/PlayerJoinListener.scala
@@ -11,7 +11,7 @@ import com.github.unchama.seichiassist.subsystems.mebius.domain.property.{
   MebiusProperty,
   NormalMebius
 }
-import com.github.unchama.seichiassist.util.{SendMessageEffect, Util}
+import com.github.unchama.seichiassist.util.{SendMessageEffect, SendSoundEffect}
 import com.github.unchama.targetedeffect.player.FocusedSoundEffect
 import net.coreprotect.model.Config
 import org.bukkit.ChatColor._
@@ -102,7 +102,7 @@ class PlayerJoinListener extends Listener {
       SendMessageEffect.sendMessageToEveryoneIgnoringPreference(
         s"${WHITE}webサイトはもう読みましたか？→$YELLOW${UNDERLINE}https://www.seichi.network/gigantic"
       )
-      Util.sendEverySound(Sound.ENTITY_PLAYER_LEVELUP, 1f, 1f)
+      SendSoundEffect.sendEverySound(Sound.ENTITY_PLAYER_LEVELUP, 1f, 1f)
 
       // ルール熟読をタイトルとチャットで迫る
       // サブタイトルと分ける理由はGUIサイズによって見切れる可能性があるため
